Fixes a typo of metric name for ifHCInOctets (#21)
* Fixes a typo of metric name for ifHCInOctets

Resolves issue #20.

* Adds Go module support to repository

The import path for the gosnmp library needed to change, as the old one
was deprecated and no longer in use. This commit also updates the
Dockerfile to use Go 1.16 and alpine 3.15.

* Changes necessary to support Go modules

I meant for this commit to be part of the previous commit (@1def767).
The previous commit introduced go.mod and go.sum, and this commit
includes some necessary changes to support that. Namely, the gosnmp
library moved to a new repository. I also updated the Docker image to Go
1.16 and alpine 3.15.

* Fixes a Go error

https://github.com/golang/go/issues/32479

* Uses fmt.Sprintf() instead of rune() for strconv