Renames metrics.yaml to metrics.yaml.sample (#23)
There was confusion (on my own part) about where the configuration file
for disco was located, as it runs on the platform cluster. The file in
this repo is/was nothing more than a sample, for reference. This commit
makes this more clear by putting "sample" in the file name. It also adds
some additional documentation to the README making note of this and
pointing to the actual config file in the k8s-support repo.